SH-model-python/SH/src/requirements.txt
python-multipart
flask
numpy
onnxruntime
torch
torchvision
pillow
flask-cors



